+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/AuxiliaryFunctions.kt
@@ -5,14 +5,8 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
-import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonColors
-import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
@@ -21,7 +15,6 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalConfiguration
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 fun resetButtons(showA: MutableState<Boolean>, showP: MutableState<Boolean>, showD: MutableState<Boolean>) {
     showA.value = true
@@ -29,11 +22,11 @@ fun resetButtons(showA: MutableState<Boolean>, showP: MutableState<Boolean>, sho
     showD.value = true
 }
 @Composable
-fun okButton(selectedItem: String, showMenu: MutableState<Boolean>) {
+fun okButton(ok: MutableState<Boolean>) {
     Spacer(Modifier.fillMaxHeight(50f / LocalConfiguration.current.screenHeightDp))
 
     val buttonWidthFraction = 0.3f
-    Button(onClick = { showMenu.value = !showMenu.value },
+    Button(onClick = { ok.value = !ok.value },
         modifier = Modifier.fillMaxHeight(120f / LocalConfiguration.current.screenHeightDp).fillMaxWidth(buttonWidthFraction)) {
         Text(text = stringResource(id = R.string.ok), fontSize = 20.sp)
     }
@@ -174,31 +167,5 @@ fun selectCategoryItemList(showA: MutableState<Boolean>, showP: MutableState<Boo
             Text(text = stringResource(id = R.string.datorii), fontSize = 20.sp)
         else
             Text(text = stringResource(id = R.string.prescurtareDatorii))
-    }
-}
-@Composable
-fun addOrDeleteItem(addButton: MutableState<Boolean>, deleteButton: MutableState<Boolean>) {
-    Button(
-        onClick = { if (!addButton.value) { addButton.value = true } },
-        shape = CircleShape,
-    ) {
-        Icon(
-            imageVector = Icons.Default.Add,
-            contentDescription = stringResource(id = R.string.favorite),
-            modifier = Modifier.fillMaxWidth(fraction = 30F / LocalConfiguration.current.screenWidthDp)
-                .fillMaxHeight(fraction = 40F / LocalConfiguration.current.screenHeightDp)
-        )
-    }
-    Spacer(modifier = Modifier.fillMaxWidth(fraction = 80F / LocalConfiguration.current.screenWidthDp))
-    Button(
-        onClick = { if (!deleteButton.value) { deleteButton.value = true } },
-        shape = CircleShape,
-    ) {
-        Icon(
-            imageVector = Icons.Default.Delete,
-            contentDescription = stringResource(id = R.string.favorite),
-            modifier = Modifier.fillMaxWidth(fraction = 30F / LocalConfiguration.current.screenWidthDp)
-                .fillMaxHeight(fraction = 40F / LocalConfiguration.current.screenHeightDp)
-        )
     }
 }
