+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/MainActivity.kt
@@ -4,12 +4,13 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import com.barbuceanuconstantin.proiectlicenta.ui.theme.ProiectLicentaTheme
+import com.barbuceanuconstantin.proiectlicenta.view.screen.showTransactionDialog
 import com.barbuceanuconstantin.proiectlicenta.view.screen.transactionsComposableScreen
 import com.barbuceanuconstantin.proiectlicenta.view.screenmodules.screen1
 import com.barbuceanuconstantin.proiectlicenta.view.screenmodules.screen2
 import com.barbuceanuconstantin.proiectlicenta.view.screenmodules.showMenu
 
-//Branch ecrane-individuale
+//Branch ecrane-individuale//
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
