+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/view/screen/CategoriesComposableScreen.kt
@@ -8,26 +8,19 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Add
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonColors
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalConfiguration
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.sp
 import com.barbuceanuconstantin.proiectlicenta.R
 import com.barbuceanuconstantin.proiectlicenta.allSubcategoriesOrTransactions
 import com.barbuceanuconstantin.proiectlicenta.data.model.Subcategory
 import com.barbuceanuconstantin.proiectlicenta.selectCategoryItemList
 import com.barbuceanuconstantin.proiectlicenta.data.model.subcategorysLazyColumn
-import com.barbuceanuconstantin.proiectlicenta.view.screenmodules.showCategoryDialog
 
 @Composable
 private fun showAddSubcategoryDialog(
@@ -35,7 +28,6 @@ private fun showAddSubcategoryDialog(
     lSP: MutableList<Subcategory>,
     lSD: MutableList<Subcategory>,
     addButton: MutableState<Boolean>,
-    deleteButton: MutableState<Boolean>,
     onDismissRequest: () -> Unit = { addButton.value = false },
     onConfirmation: () -> Unit = { addButton.value = false },
 ) {
@@ -51,10 +43,9 @@ private fun showAddSubcategoryDialog(
 
 @Composable
 fun categoriesComposableScreen(showA: MutableState<Boolean>, showP: MutableState<Boolean>, showD: MutableState<Boolean>,
-                     addButton: MutableState<Boolean>, deleteButton: MutableState<Boolean>,
-                     lSA: MutableList<Subcategory>, lSP: MutableList<Subcategory>, lSD: MutableList<Subcategory>) {
+                     addButton: MutableState<Boolean>, lSA: MutableList<Subcategory>, lSP: MutableList<Subcategory>, lSD: MutableList<Subcategory>) {
     if (addButton.value) {
-        showAddSubcategoryDialog(lSA = lSA, lSP = lSP, lSD = lSD, addButton = addButton, deleteButton = deleteButton)
+        showAddSubcategoryDialog(lSA = lSA, lSP = lSP, lSD = lSD, addButton = addButton)
     }
     if (!addButton.value) {
         Scaffold(
