+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/view/screen/CategoryModifyDialogScreen.kt
@@ -1,4 +1,4 @@
-package com.barbuceanuconstantin.proiectlicenta.view.screenmodules
+package com.barbuceanuconstantin.proiectlicenta.view.screen
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -6,7 +6,6 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.icons.Icons
