+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/view/screenmodules/CurrenciesMenu.kt
@@ -82,7 +82,7 @@ class MeniuValute {
                 }
             }
 
-            okButton(selectedItem, showMeniuValute)
+            okButton(showMeniuValute)
         }
     }
 }
