+++ app/src/main/java/com/barbuceanuconstantin/proiectlicenta/view/screenmodules/SubcategoriesMenu.kt
@@ -71,7 +71,7 @@ class MeniuSubcategorys {
                     )
                 }
             }
-            okButton(selectedItem, showMeniuSubcategorys)
+            okButton(showMeniuSubcategorys)
         }
     }
 }
